+++ tests/TestCase.php
@@ -3,7 +3,9 @@
 namespace Convenia\Pigeon\Tests;
 
 use Mockery;
+use ArrayAccess;
 use Orchestra\Testbench\TestCase as BaseTestCase;
+use PHPUnit\Util\InvalidArgumentHelper;
 
 class TestCase extends BaseTestCase
 {