+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `on-failure` config to ack, reject or throw exception
 - Added `Pigeon::headers([])` and config `headers` key
+- Added possibility to use callable on headers config
 
 ## [1.2.0]
 ### Added